Add Bookmarks & Auto-Bookmarking
Bookmarks are tracked on the session, and must be opted in per event by
using the `bookmark` tag. There is also a new Bookmarks button in the
header to show the list of bookmarks in a modal.

Bookmarks have a "type" which determines how it is displayed in the
modal list. See the constant bookmarkTypeMapping for reference.

Auto-bookmarking is done as an additional action added to the Moment a
`bond`, `unbond`, or `progress` tag are used. For `progress`, it uses
the third argument to create additional bookmark types that indicate
whether it's been newly started, in progress, or complete.